Renamed ReactProfileTimer -> ReactProfilerTimer
